Move state-derived props except for `guess` in `LanguageDropdown`
